Install frontend assets only if requested
Add the variable `frontend_include_branding`, which governs whether the
private dpla_frontend_assets gem is included for branding-related
assets.  This defuaults to false.

Addresses one of the problems raised in
https://github.com/dpla/automation/issues/117

+++ ansible/roles/frontend/files/build_frontend.sh
@@ -1,16 +1,44 @@
 #!/bin/bash
 
+# Synopsis
+#
+# - Build with Ruby 1.9.3-p547, with dpla_frontend_assets gem
+#   build_frontend.sh 1.9.3-p547
+#
+# - Build as above, but without dpla_frontend_assets gem
+#   build_frontend.sh 1.9.3-p547 nobranding
+
+# Known issue:
+#
+# If you run `bundle install --with dpla_branding` and then do
+# `bundle install --without dpla_branding`, Bundler will keep trying to fetch
+# the dpla_frontend_assets gem when Unicorn is run with `bundle exec`, which
+# will result in errors.  We don't know of a way around this yet.  It works fine
+# if you start with `--without` and later switch to `--with`.
+
 USE_VERSION=$1
+NO_BRANDING=${2:-""}
 export PATH=$HOME/.rbenv/bin:$PATH
 LOGFILE=/tmp/build_frontend.log
+
+nobranding() {
+    if [ "$NO_BRANDING" == "nobranding" ]; then
+        return 0
+    else
+        return 1
+    fi
+}
 
 echo "starting" > $LOGFILE
 
 eval "`rbenv init -`" >> $LOGFILE 2>&1
-# Start ssh-agent and set environment variables.
-# Work-around for private GitHub repository in Gemfile.
-eval `ssh-agent`
-ssh-add $HOME/git_private_key  >> $LOGFILE 2>&1
+
+if ! nobranding; then
+    # Start ssh-agent and set environment variables.
+    # Work-around for private GitHub repository in Gemfile.
+    eval `ssh-agent`
+    ssh-add $HOME/git_private_key  >> $LOGFILE 2>&1
+fi
 
 cd $HOME/frontend
 
@@ -19,16 +47,23 @@ rbenv shell $USE_VERSION >> $LOGFILE 2>&1
 echo "installing bundle ..." >> $LOGFILE
 
 rm -f Gemfile.lock
-bundle install >> $LOGFILE 2>&1
-bundle update dpla_frontend_assets >> $LOGFILE 2>&1
+if nobranding; then
+    install_opts="--without dpla_branding"
+else
+    install_opts="--with dpla_branding"
+fi
+echo "install_opts: $install_opts" >> $LOGFILE
+bundle install $install_opts >> $LOGFILE 2>&1
 rbenv rehash
 
 echo "precompiling assets ..." >> $LOGFILE
 bundle exec rake assets:precompile >> $LOGFILE 2>&1
 
-echo "killing ssh_agent ..." >> $LOGFILE
-# Variable set above by ssh-agent
-kill $SSH_AGENT_PID
+if ! nobranding; then
+    echo "killing ssh_agent ..." >> $LOGFILE
+    # Variable set above by ssh-agent
+    kill $SSH_AGENT_PID
+fi
 
 echo "rsync home to /srv/www ..." >> $LOGFILE
 /usr/bin/rsync -rIptogl --checksum --delete --delay-updates \
